next changes

+++ client/src/store/Reports/components/ReportsListContainer.jsx
@@ -29,19 +29,21 @@ const ReportsListContainer = () => {
     const loggedUser = useSelector(state => state.users.logged_user);
     const reportsList = useSelector(state => state.reports.reports);
 
-    console.log('reports list ...');
-
     const [ message, setMessage ] = useState(null);
     const [ month, setMonth ] = useState(presentMonth());
-    const [ reports, setReports ] = useState([]);
+    // const [ reports, setReports ] = useState([]);
     const [ filteredReports, setFilteredReports ] = useState([]);
 
     const dispatch = useDispatch();
 
-    console.log('filteredReports',filteredReports);
+    useEffect(() => {
+        setTimeout(() => {
+            setMessage(null);
+        },5000);
+    },[message]);
     
     useEffect(() => {
-        setReports(reportsList);
+        // setReports(reportsList);
         const selected = Reportselector();
         setFilteredReports(selected);
     },[reportsList]);
@@ -49,7 +51,7 @@ const ReportsListContainer = () => {
     useEffect(()=>{
         const selected = Reportselector(reportsList);
         setFilteredReports(selected);
-    },[month])
+    },[month]);
 
     const Reportselector = () => {
         const selected = reportsList.filter(row => {
@@ -72,14 +74,16 @@ const ReportsListContainer = () => {
                 Marian: report.Marian,
                 Piotrek: report.Piotrek
             }
-
             dispatch(addReport(data));
-
+            setMessage('Rekord został dodany');
         }
     }
 
     const deleteHandler = (id) => {
-        dispatch(removeReport(id));
+        if(window.confirm("Jeśli chcesz usunąć wybierz tak") === true){
+            dispatch(removeReport(id));
+            setMessage('Rekord został usunięty');
+        }
     }
 
     const changeMonthHandler = (event) => {
@@ -97,16 +101,18 @@ const ReportsListContainer = () => {
                 <tr className={styles.tr}><th>Nr</th><th>Dzień</th><th>Opis</th><th>Marian godz.</th><th>Piotrek godz.</th></tr>
                 { list }
             </tbody> 
-        </table> : 'No results found';
+        </table> : 'Brak wyników dla wybranego miesiąca';
 
     return(
-        <>
-            <div className="message">{ message }</div>
-            <ReportsMonthSelector selectedMonth={month} changeMonthHandler={changeMonthHandler} />
+        <div className={styles["reports-list-container-box"]}>
+            <header>
+                <ReportsMonthSelector selectedMonth={month} changeMonthHandler={changeMonthHandler} />
+                { message && <div className={styles.message}>{ message }</div> }
+            </header>
             <ReportsAddForm submitHandler={submitHandler} />
             <div className={styles['report-items-box']}>{ content }</div>
             <ReportsSummary MarianSalary={salary[0].Marian} PiotrekSalary={salary[0].Piotrek} reports={filteredReports} />
-        </>
+        </div>
     )
 }
